refactor(databese): made services communicate by injection
