+++ projects/core/src/lib/authorization/components/sign-in/sign-in.component.ts
@@ -5,7 +5,7 @@ import { User } from "../../models/user";
 import { Subject, takeUntil } from "rxjs";
 import { Routes } from "../../models/routes";
 import { Roles } from "../../models/roles";
-import { UsersDataService } from "../../services/data/users.data.service";
+import { UsersDataService } from "../../../services/users.data.service";
 
 @Component({
 	selector: "jsmu-sign-in",
